Implements new range mechanics for #690.

+++ src/main/java/org/lisoft/lsml/model/modifiers/Attribute.java
@@ -115,7 +115,7 @@ public class Attribute {
      * @param aAmount
      *            The new base value.
      */
-    public void setBaseValue(int aAmount) {
+    public void setBaseValue(double aAmount) {
         baseValue = aAmount;
     }
